Mv creating setup.py to init
