+++ tests/image/test_build.py
@@ -19,14 +19,11 @@ def test_cli(docker_client_mock, project_dir, test_mock):
         {"stream": "all is ok"},
         {"stream": "Successfully built 12345"}
     ]
-    setup_py_path = os.path.join(project_dir, 'setup.py')
-    assert not os.path.isfile(setup_py_path)
     runner = CliRunner()
     result = runner.invoke(cli, ["dev", "-v"])
     assert result.exit_code == 0
     docker_client_mock.build.assert_called_with(
         decode=True, path=project_dir, tag='registry/user/project:1.0')
-    assert os.path.isfile(setup_py_path)
     test_mock.assert_called_with("dev", None)
 
 
@@ -53,6 +50,7 @@ def test_cli_no_dockerfile(docker_client_mock, project_dir):
     result = runner.invoke(cli, ["dev"])
     assert result.exit_code == shub_exceptions.BadParameterException.exit_code
 
+
 @pytest.mark.usefixtures('project_dir')
 def test_cli_fail(docker_client_mock):
     docker_client_mock.build.return_value = [
@@ -69,12 +67,9 @@ def test_cli_skip_tests(docker_client_mock, test_mock, project_dir, skip_tests_f
         {"stream": "all is ok"},
         {"stream": "Successfully built 12345"}
     ]
-    setup_py_path = os.path.join(project_dir, 'setup.py')
-    assert not os.path.isfile(setup_py_path)
     runner = CliRunner()
     result = runner.invoke(cli, ["dev", skip_tests_flag])
     assert result.exit_code == 0
     docker_client_mock.build.assert_called_with(
         decode=True, path=project_dir, tag='registry/user/project:1.0')
-    assert os.path.isfile(setup_py_path)
     assert test_mock.call_count == 0
